requirements.txt
Pillow
django-widget-tweaks
psycopg2
